+++ backend/integration_tests/src/client/group_index.rs
@@ -22,6 +22,7 @@ generate_update_call!(set_group_verification);
 
 pub mod happy_path {
     use candid::Principal;
+    use constants::CHUNK_STORE_CHUNK_SIZE;
     use group_index_canister::ChildCanisterType;
     use pocket_ic::PocketIc;
     use sha256::sha256;
@@ -226,7 +227,7 @@ pub mod happy_path {
         wasm: &[u8],
         canister_type: ChildCanisterType,
     ) {
-        for (index, chunk) in wasm.chunks(1_000_000).enumerate() {
+        for (index, chunk) in wasm.chunks(CHUNK_STORE_CHUNK_SIZE).enumerate() {
             let response = super::upload_wasm_chunk(
                 env,
                 sender,