+++ backend/canisters/community/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Allow community owners to delete channels and explore private channels ([#7652](https://github.com/open-chat-labs/open-chat/pull/7652))
+
 ## [[2.0.1657](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1657-community)] - 2025-03-17
 
 ### Added